+++ packages/subgraph/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- Remove `chainId` from fetching deposit events
 - Update deposit manager address
 - Remove `wallet` in place of `handler` contract for syncing sdk events
 - Add entities/events/handler for DepositManager `DepositInstantiated` event